+++ ThirdPartyAdapters/chartboost/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Chartboost Android Mediation Adapter Changelog
 
+#### Next Version
+- Updated the adapter to use new `AdError` API.
+
 #### Version 8.3.0.0
 - Verified compatibility with Chartboost SDK 8.3.0.
 - Updated the minimum required Google Mobile Ads SDK version to 20.5.0.